refactor: pretty to tree
